Renamed MemoryMapped project - added .Queue to name and path
