+++ src/Serilog.MemoryMapped.Queue/IMemoryMappedQueue.cs
