+++ src/Serilog.MemoryMapped.Queue/MemoryMappedMonitor.cs
