+++ src/Serilog.MemoryMapped.Queue/MemoryMappedQueue.cs
