+++ src/Serilog.MemoryMapped.Queue/MemoryMappedQueueBuffer.cs
