+++ src/Serilog.MemoryMapped.Queue/MemoryMappedQueueStats.cs
