+++ src/Serilog.MemoryMapped.Queue/MemoryMapperLogger.cs
