+++ src/Serilog.MemoryMapped.Queue/NameNormalizer.cs
