+++ src/Serilog.MemoryMapped.Queue/PooledBufferWriter.cs
