+++ src/Serilog.MemoryMapped.Queue/Serializers/IFastSerializer.cs
