+++ src/Serilog.MemoryMapped.Queue/Serilog.MemoryMapped.Queue.csproj
